+++ packages/message-manager/src/EncryptionPublicKeyManager.ts
@@ -1,9 +1,9 @@
 import type {
-  ActionConstraint,
-  EventConstraint,
-  RestrictedMessenger,
-} from '@metamask/base-controller';
+  ControllerGetStateAction,
+  ControllerStateChangeEvent,
+} from '@metamask/base-controller/next';
 import { ApprovalType } from '@metamask/controller-utils';
+import type { Messenger } from '@metamask/messenger';
 
 import type {
   AbstractMessage,
@@ -31,14 +31,23 @@ export type EncryptionPublicKeyManagerUpdateBadgeEvent = {
   payload: [];
 };
 
-export type EncryptionPublicKeyManagerMessenger = RestrictedMessenger<
+type EncryptionPublicKeyManagerActions = ControllerGetStateAction<
   typeof managerName,
-  ActionConstraint,
-  | EventConstraint
+  EncryptionPublicKeyManagerState
+>;
+
+type EncryptionPublicKeyManagerEvents =
+  | ControllerStateChangeEvent<
+      typeof managerName,
+      EncryptionPublicKeyManagerState
+    >
   | EncryptionPublicKeyManagerUnapprovedMessageAddedEvent
-  | EncryptionPublicKeyManagerUpdateBadgeEvent,
-  string,
-  string
+  | EncryptionPublicKeyManagerUpdateBadgeEvent;
+
+export type EncryptionPublicKeyManagerMessenger = Messenger<
+  typeof managerName,
+  EncryptionPublicKeyManagerActions,
+  EncryptionPublicKeyManagerEvents
 >;
 
 type EncryptionPublicKeyManagerOptions = {
@@ -95,10 +104,7 @@ export class EncryptionPublicKeyManager extends AbstractMessageManager<
   EncryptionPublicKey,
   EncryptionPublicKeyParams,
   EncryptionPublicKeyParamsMetamask,
-  ActionConstraint,
-  | EventConstraint
-  | EncryptionPublicKeyManagerUnapprovedMessageAddedEvent
-  | EncryptionPublicKeyManagerUpdateBadgeEvent
+  EncryptionPublicKeyManagerMessenger
 > {
   constructor({
     additionalFinishStatuses,
@@ -185,7 +191,7 @@ export class EncryptionPublicKeyManager extends AbstractMessageManager<
     const messageId = messageData.id;
 
     await this.addMessage(messageData);
-    this.messagingSystem.publish(`${this.name}:unapprovedMessage`, {
+    this.messenger.publish(`${this.name}:unapprovedMessage` as const, {
       ...updatedMessageParams,
       metamaskId: messageId,
     });