+++ packages/notification-services-controller/src/NotificationServicesPushController/__fixtures__/mockMessenger.ts
@@ -1,21 +1,57 @@
-import { Messenger } from '@metamask/base-controller';
+import {
+  Messenger,
+  MOCK_ANY_NAMESPACE,
+  type MessengerActions,
+  type MessengerEvents,
+  type MockAnyNamespace,
+} from '@metamask/messenger';
 
-import type {
-  AllowedActions,
-  AllowedEvents,
-  NotificationServicesPushControllerMessenger,
-} from '..';
+import type { NotificationServicesPushControllerMessenger } from '..';
+
+const controllerName = 'NotificationServicesPushController';
+
+type AllNotificationServicesPushControllerActions =
+  MessengerActions<NotificationServicesPushControllerMessenger>;
+
+type AllNotificationServicesPushControllerEvents =
+  MessengerEvents<NotificationServicesPushControllerMessenger>;
+
+type RootMessenger = Messenger<
+  MockAnyNamespace,
+  AllNotificationServicesPushControllerActions,
+  AllNotificationServicesPushControllerEvents
+>;
+
+/**
+ * Creates and returns a root messenger for testing
+ *
+ * @returns A messenger instance
+ */
+function getRootMessenger(): RootMessenger {
+  return new Messenger({
+    namespace: MOCK_ANY_NAMESPACE,
+  });
+}
 
 export const buildPushPlatformNotificationsControllerMessenger =
   (): NotificationServicesPushControllerMessenger => {
-    const globalMessenger = new Messenger<AllowedActions, AllowedEvents>();
+    const rootMessenger = getRootMessenger();
 
-    return globalMessenger.getRestricted<
-      'NotificationServicesPushController',
-      AllowedActions['type']
+    const messenger = new Messenger<
+      typeof controllerName,
+      AllNotificationServicesPushControllerActions,
+      AllNotificationServicesPushControllerEvents,
+      RootMessenger
     >({
-      name: 'NotificationServicesPushController',
-      allowedActions: ['AuthenticationController:getBearerToken'],
-      allowedEvents: [],
+      namespace: controllerName,
+      parent: rootMessenger,
     });
+
+    rootMessenger.delegate({
+      messenger,
+      actions: ['AuthenticationController:getBearerToken'],
+      events: [],
+    });
+
+    return messenger;
   };