+++ packages/remote-feature-flag-controller/src/index.ts
@@ -1,5 +1,6 @@
 export { RemoteFeatureFlagController } from './remote-feature-flag-controller';
 export type {
+  RemoteFeatureFlagControllerState,
   RemoteFeatureFlagControllerMessenger,
   RemoteFeatureFlagControllerActions,
   RemoteFeatureFlagControllerGetStateAction,
@@ -13,9 +14,6 @@ export {
   EnvironmentType,
 } from './remote-feature-flag-controller-types';
 
-export type {
-  RemoteFeatureFlagControllerState,
-  FeatureFlags,
-} from './remote-feature-flag-controller-types';
+export type { FeatureFlags } from './remote-feature-flag-controller-types';
 export { ClientConfigApiService } from './client-config-api-service/client-config-api-service';
 export { generateDeterministicRandomNumber } from './utils/user-segmentation-utils';