+++ packages/shield-controller/src/index.ts
@@ -9,6 +9,7 @@ export type {
   ShieldControllerEvents,
   ShieldControllerMessenger,
   ShieldControllerState,
+  ShieldControllerGetStateAction,
   ShieldControllerCheckCoverageAction,
   ShieldControllerCoverageResultReceivedEvent,
   ShieldControllerStateChangeEvent,